requirements.txt
streamlit
pytesseract
pandas
openpyxl
pillow
pymupdf
xlsxwriter
numpy
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
opencv-python-headless
